Added extra stability for loaded images
A bug cropped up when the plugin was called on images that were already
loaded or cached.

Added a section to detect if an image has a width set, in which case is
most likely already loaded.

When this ‘width’ test passes the plugin skips the `$image.load()` call
and begins manipulating the image straight away.